requirements.txt
Flask
flask_sqlalchemy
flask_admin
python-dotenv
psycopg2-binary
line-bot-sdk==2.4.1
pytz
pillow
pytesseract
opencv-python-headless
numpy
gunicorn
wtforms==2.3.3
redis
gspread
oauth2client
Flask-Migrate
Flask-WTF